requirements.txt
Django==2.0.7
django-memoize==2.1.1
djangorestframework==3.8.2
requests==2.19.1
